refactor(meeting): DTO 가져오기 경로 업데이트
